requirement.txt
streamlit
google-generativeai
pdfplumber
python-dotenv
pandas
